[UPDATE] Refactorizando PostItem to DeprecatedPostItem component y actualizando imports
